# [NVRHI] Build-time performance improvement for precompiled headers