--- a/neo/renderer/GuiModel.h
+++ b/neo/renderer/GuiModel.h
@@ -43,10 +43,11 @@ struct guiModelSurface_t
 class idRenderMatrix;
 class Framebuffer;
 
-namespace ImGui
-{
-struct ImDrawData;
-}
+// SRS - not needed, causes mismatched types when calling ImGui_RenderDrawLists( ImDrawData* )
+//namespace ImGui
+//{
+//struct ImDrawData;
+//}
 
 struct ImDrawData;