--- a/neo/renderer/RenderCommon.h
+++ b/neo/renderer/RenderCommon.h
@@ -40,27 +40,6 @@ If you have questions concerning this license or the applicable additional terms
 #include "Font.h"
 #include "Framebuffer.h"
 
-#if defined( USE_NVRHI )
-
-	#if USE_DX11 || USE_DX12
-		#include <DXGI.h>
-	#endif
-
-	#if USE_DX11
-		#include <d3d11.h>
-	#endif
-
-	#if USE_DX12
-		#include <d3d12.h>
-	#endif
-
-	#if USE_VK
-		#include <nvrhi/vulkan.h>
-	#endif
-
-	#include <nvrhi/nvrhi.h>
-#endif
-
 // maximum texture units
 const int MAX_PROG_TEXTURE_PARMS	= 16;
 
@@ -1415,15 +1394,14 @@ struct glimpParms_t
 
 // Eric: If on Linux using Vulkan use the sdl_vkimp.cpp methods
 // SRS - Generalized Vulkan SDL platform
-#if defined(VULKAN_USE_PLATFORM_SDL)
+#if defined( VULKAN_USE_PLATFORM_SDL )
 #include <vector>
 
 #define CLAMP(x, lo, hi)    ((x) < (lo) ? (lo) : (x) > (hi) ? (hi) : (x))
-// Helper functions for using SDL2 and Vulkan on Linux.
+// Helper function for using SDL2 and Vulkan on Linux.
 std::vector<const char*> get_required_extensions();
-#if defined( USE_NVRHI )
-	vk::Result CreateSDLWindowSurface( vk::Instance instance, vk::SurfaceKHR* surface );
-#else
+
+#if !defined( USE_NVRHI )
 	extern vulkanContext_t vkcontext;
 #endif