--- a/neo/sys/DeviceManager_VK.cpp
+++ b/neo/sys/DeviceManager_VK.cpp
@@ -27,6 +27,7 @@
 #include <queue>
 #include <unordered_set>
 
+#include "renderer/RenderCommon.h"
 #include <sys/DeviceManager.h>
 
 #include <nvrhi/vulkan.h>