--- a/neo/sys/sdl/sdl_events.cpp
+++ b/neo/sys/sdl/sdl_events.cpp
@@ -853,7 +853,7 @@ void Sys_GrabMouseCursor( bool grabIt )
 		flags = GRAB_SETSTATE;
 	}
 // SRS - Generalized Vulkan SDL platform
-#if defined(VULKAN_USE_PLATFORM_SDL)
+#if defined( VULKAN_USE_PLATFORM_SDL )
 	VKimp_GrabInput( flags );
 #else
 	GLimp_GrabInput( flags );