--- a/neo/ui/DeviceContext.cpp
+++ b/neo/ui/DeviceContext.cpp
@@ -32,7 +32,7 @@ If you have questions concerning this license or the applicable additional terms
 #include "DeviceContext.h"
 
 #include "libs/imgui/imgui.h"
-#include "../renderer/GuiModel.h"
+#include "../renderer/RenderCommon.h"
 
 extern idCVar in_useJoystick;