**UPDATE: retargeted to master now that 635-nvrhi5 has been merged, results should be similar**

Here is a small present for the season:

I noticed that PCH build performance was limited in the NVRHI stream. Having a deeper look showed that `precompiled.h` could be improved by adding `RenderCommon.h`, however some restructuring of renderer-specific library includes (e.g. `d3d12.h`, `nvrhi/vulkan.h`, etc) would be needed. A positive side-effect is that it cleans up and restricts the location of the **USE_DX12**, **USE_VK**, etc. defines and related includes to `DeviceManager.h` and `DeviceManager_<renderer>.cpp`.

Here is the result of that work and the results are significant. The data below are fresh builds including shader compile time:

Windows 10 build using VS2019:
Before: 97 seconds
After: **52 seconds**

Linux (Manjaro) build using Code:Blocks/clang**:
Before: 113 seconds
After: **74 seconds**

macOS Monterey build using Xcode 13/clang:
Before: 120 seconds
After: **63 seconds**

** Note: Linux timings were measured when booting off a USB stick, so absolute results may be better on a real SSD

Overall the results look good, with an almost 2x improvement in build time for Windows and macOS. Linux data is TBD pending verification on a machine with an SSD build drive.

Also fixes a minor `c_str()` issue inside `idCameraAnim::gltfLoadAnim()`, otherwise clang complains.

Merry Christmas to all!

**UPDATE**: Simplify **precompiled.h.gch** cleanup using cmake custom command vs. a custom target.  This makes the code simpler and works more reliably cross platform for linux/macOS.