@SRSaunders 2023-01-14
Hi Robert - I'm glad you found this useful and were able to verify the build-time improvement on your setup.

I currently have two other pending pull requests:

1. #736 which is important for linux/macOS since it fixes a bad regression in the posix basepath logic (from one of my previous PRs - sorry about that) which can prevent writing of generated lighting files (or any other generated files).  It also fixes a few other things (SDL mode filtering & window handling, some linux/macOS-specifics, and a few generic items).
2. #645 is a long-standing one regarding Cinematics.  While it was labelled initially as a feature submission, it is primarily a set of fixes and memory management improvements for cinematics and ffmpeg.  It now also contains changes that re-enable the testVideo command for the Vulkan platform, which was broken by the migration to NVRHI.

If you need any additional testing or changes before you consider them for merging, please let me know.